mlops_pipeline/requirements_mlops.txt
# MLOps Pipeline Requirements
# Core ML Libraries
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# MLOps and Experiment Tracking
mlflow>=2.8.0
wandb>=0.16.0
optuna>=3.4.0

# Model Serving and API
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
requests>=2.31.0

# Model Monitoring and Drift Detection
evidently>=0.4.0
alibi-detect>=0.11.0
great-expectations>=0.18.0

# Data Engineering and Pipelines
apache-airflow>=2.7.0
dask>=2023.10.0
prefect>=2.14.0

# Feature Store
feast>=0.34.0

# Database and Storage
sqlalchemy>=2.0.0
redis>=5.0.0
psycopg2-binary>=2.9.0

# Cloud and Deployment
docker>=6.1.0
kubernetes>=28.1.0
boto3>=1.34.0
azure-storage-blob>=12.19.0

# Monitoring and Observability
prometheus-client>=0.19.0
grafana-api>=1.0.0
elasticsearch>=8.11.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.1.0
mypy>=1.7.0

# Documentation
mkdocs>=1.5.0
mkdocs-material>=9.4.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.1
click>=8.1.0
tqdm>=4.66.0
joblib>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0
streamlit>=1.28.0

# Time Series and Advanced Analytics
statsmodels>=0.14.0
prophet>=1.1.4
tsfresh>=0.20.0
